% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_is_multipanelfigure}
\alias{assert_is_multipanelfigure}
\title{Check that the input is a multipanelfigure}
\usage{
assert_is_multipanelfigure(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Checks that the input is of class \code{multipanelfigure} and has the
appropriate attributes.
}
\references{
Graumann, J., and Cotton, R.J. (2018). multipanelfigure: Simple
Assembly of Multiple Plots and Images into a Compound Figure. Journal of
Statistical Software 84. doi: 10.18637/jss.v084.c03
}
