% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffsave.R
\name{ffsave}
\alias{ffsave}
\title{Save a multi-panel figure frame}
\usage{
ffsave(filename = default_name(plot), figureFrame,
  device = pathological::get_extension(filename), path = NULL, dpi = 300,
  ...)
}
\arguments{
\item{filename}{A string specifying the filename, or a path to the file.}

\item{figureFrame}{An object created by \code{multipanelfigure}.}

\item{device}{Either a device function, or a string naming the device.
Automatically determined from the \code{filename}.}

\item{path}{A string naming the directoy to store the output file (for use
when \code{filename} is just the name of the file.)}

\item{dpi}{The resolution of the output, in dots per inch.  (Only used by
bitmap devices.)}

\item{...}{Passed to the device function.}
}
\value{
The filename is invisibly returned, but the function is mostly
invoked for the side-effect of writing a plot to file.
}
\description{
Saves a multi-panel figure frame to a file, using sensible defaults.
}
\references{
This function is derived from \code{\link[ggplot2]{ggsave}}.
}

