% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-docs.R
\name{multipanelfigure-deprecated}
\alias{multipanelfigure-deprecated}
\title{Superceded objects in multipanelfigure}
\usage{
addPanel( ... )
}
\arguments{
\item{...}{Arguments to functions you shouldn't use.}
}
\description{
Functions that are no longer used or have been superceded by
functions with underscore-separated names.
}

