% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipanelfigure.R
\name{multipanelfigure}
\alias{multipanelfigure}
\title{multipanelfigure}
\usage{
multipanelfigure(width = NULL, widths = NULL, columns = NULL,
  height = NULL, heights = NULL, rows = NULL, interPanelSpacing = NaN,
  unit = "mm", figureName = "FigureX", panelLabels = LETTERS)
}
\arguments{
\item{width}{Single \code{link{numeric}} defining the width of the resulting
\code{\link[gtable]{gtable}} (unit depending on \code{unit}). See 'Details' for
dependend and interfering parameters.}

\item{widths}{\code{\link{vector}} of \code{\link{numeric}}s defining the
actual widths of panels/columns in the resulting \code{\link[gtable]{gtable}} (unit
depending on \code{unit}). See 'Details' for dependend and
interfering parameters.}

\item{columns}{Single \code{\link{numeric}} defining the number of columns in
the resulting \code{\link[gtable]{gtable}}. See 'Details' for dependend and
interfering parameters.}

\item{height}{Single \code{link{numeric}} defining the height of the resulting
\code{\link[gtable]{gtable}} (unit depending on \code{unit}). See 'Details' for
dependend and interfering parameters.}

\item{heights}{\code{\link{vector}} of \code{\link{numeric}}s defining the
actual heights of panels/rows in the resulting \code{\link[gtable]{gtable}} (unit
depending on \code{unit}). See 'Details' for dependend and
interfering parameters.}

\item{rows}{Single \code{\link{numeric}} defining the number of rows in
the resulting \code{\link[gtable]{gtable}}. See 'Details' for dependend and
interfering parameters.}

\item{interPanelSpacing}{The amount of white space automatically inserted
between panels. Defaults to \code{5 mm} unless explicitly given, in which
case the value depents on the \code{unit} parameter.}

\item{unit}{Single \code{\link{character}} object defining the unit of all
dimensions defined. Must satisfy \code{grid:::valid.units}.}

\item{figureName}{Single \code{\link{character}} object defining the name of
the resulting \code{\link[gtable]{gtable}}.}

\item{panelLabels}{A \code{\link{character}} \code{\link{vector}} defining
the panel labels used for automated annotation. \code{\link{length}} must be
larger or equal to the number of panels defined. Will be used sequentially.}
}
\value{
Returns an object of class \code{multipanelfigure} as well as
\code{\link[gtable]{gtable}} object with the following additional attributes:
\describe{
  \item{\code{multipanelfigure.panelsFree}:}{A \code{\link{logical}}
    \code{\link{matrix}} with the dimensions of the \code{\link[gtable]{gtable}}
    indicating occupancy of the panels in the table.}
  \item{\code{multipanelfigure.panelLabelsFree}:}{A \code{\link{character}}
    \code{\link{vector}} indicative of the \code{panelLabels} still available.}
  \item{\code{multipanelfigure.unit}:}{A single \code{\link{character}}
    object storing the corresponding value given during object creation.}}
}
\description{
A convenience function building \code{\link[gtable]{gtable}}-based
infrastructure for the assembly of multipanel figures.
}
\details{
The \code{\link[gtable]{gtable}} may be constructed in two ways:
\enumerate{
  \item{Based on explicit width/heigth definitions for individual panels
    handed to it via \code{widths} and \code{heights}.}
  \item{Based on total figure/\code{\link[gtable]{gtable}} dimensions given by
    \code{width} and \code{height} together with the number of \code{columns}
    and \code{rows} requested.}}
The function automatically inserts whitespace of width
\code{interPanelSpacing} between panels, which has to be considered for the
total dimensions of the resulting \code{\link[gtable]{gtable}}. Width of the
\code{\link[gtable]{gtable}} in the former case, for example may be calculated
\deqn{W[total] = sum(widths) + (length(widths) - 1) * interPanelSpacing}
while width of resulting panels in the latter table construction approach may
be calculated
\deqn{W[panel] = (width - (columns - 1) * interPanelSpacing)/columns}

The two approaches to \code{\link[gtable]{gtable}} construction require mutually
exclusive parameter sets:
\describe{
  \item{Individual definition of panel dimensions:}{Requires \code{widths} and
    \code{heights}. Excludes the use of \code{width}, \code{columns},
    \code{heights} and \code{rows}.}
  \item{Definition of global \code{\link[gtable]{gtable}}/figure dimensions:}{
    Requires \code{width}, \code{columns}, \code{heights} and \code{rows}.
    Excludes the use of \code{widths} and \code{heights}.}}
}
\examples{
library(gtable)
# Figure construction based on overall dimensions
Figure1 <- multipanelfigure(
   width = 100,
   columns = 4,
   height = 100,
   rows = 6,
   figureName = "Figure1")
gtable_show_layout(Figure1)

# Figure construction based on individual panel dimensions
Figure2 <- multipanelfigure(
   widths = c(20,30),
   heights = c(40,60),
   figureName = "Figure2")
gtable_show_layout(Figure2)

# A more involved example including filling and printing to device ...
## Make a simple ggplot object to fill panels
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()
## Fill panels
Figure2 <- addpanel(Figure2, p, topPanel = 1, leftPanel = 2)
Figure2 <- addpanel(Figure2, p, topPanel = 2, leftPanel = 1, rightPanel = 2)
## Plot to appropriately sized png device
tmpFile <- tempfile(fileext = ".png")
usedUnit <- "in"
width <- simplegrobwidth(Figure2, unitTo = usedUnit)
height <- simplegrobheight(Figure2, unitTo = usedUnit)
ggsave(tmpFile, Figure2, width = width, height = height)
message(
  paste0("Now have a look at '",tmpFile,"' - nicely sized PNG output."))
}
\author{
Johannes Graumann
}
\seealso{
\code{\link[gtable]{gtable}}, \code{\link{addpanel}},
\code{\link{simplegrobwidth}}, \code{\link{simplegrobheight}}
}
\keyword{hplot}
\keyword{utilities}

