% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobdimensions.R
\name{simplegrobwidth}
\alias{simplegrobheight}
\alias{simplegrobwidth}
\title{Convenient Access to \code{grob} Dimensions}
\usage{
simplegrobwidth(grob, unitTo = "mm")
simplegrobheight(grob, unitTo = "mm")
}
\arguments{
\item{grob}{A \code{\link[grid]{grob}} object for which dimensions are to be
retrieved.}

\item{unitTo}{A single \code{\link{character}} string representing a valid
\pkg{grid}-\code{\link[grid]{unit}}.}
}
\value{
Single \code{\link{numeric}} objects are returned.
}
\description{
Convenience functions extracting dimensions from
\code{\link{grob}} objects.
}
\examples{
# Get dimensions of a grid grob
testCircle <- grid::circleGrob(x = 15, y = 30, r = 15, default.unit = "mm")
simplegrobheight(testCircle)
simplegrobwidth(testCircle)

# Use the unitTo arg to convert units
simplegrobheight(testCircle, unitTo = "in")
simplegrobwidth(testCircle, unitTo = "cm")

# Get dimensions of a multi-panel figure
Figure <- multipanelfigure(width = 55, height = 55, rows = 2, columns = 2)
simplegrobheight(Figure)
simplegrobwidth(Figure)

# ggsave defaults to measuring dimensions in inches
usedUnits <- "in"
width <- simplegrobwidth(Figure, unitTo = usedUnits)
height <- simplegrobheight(Figure, unitTo = usedUnits)
tmpFile <- tempfile(fileext = ".png")
ggplot2::ggsave(tmpFile, Figure, width = width, height = height)
\donttest{ # Not testing due to use of external software
utils::browseURL(tmpFile)
}
}
\author{
Johannes Graumann
}
\seealso{
\code{\link{multipanelfigure}}
}

