% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_dimensions.R
\name{grob_dimensions}
\alias{grob_dimensions}
\alias{simple_grob_height}
\alias{simple_grob_width}
\alias{simplegrobheight}
\alias{simplegrobwidth}
\title{Convenient Access to \code{grob} Dimensions}
\usage{
simple_grob_width(grob, unitTo = "mm")
simple_grob_height(grob, unitTo = "mm")
}
\arguments{
\item{grob}{A \code{\link[grid]{grob}} object for which dimensions are to be
retrieved.}

\item{unitTo}{A single \code{\link{character}} string representing a valid
\pkg{grid}-\code{\link[grid]{unit}}.}
}
\value{
Single \code{\link{numeric}} objects are returned.
}
\description{
Convenience functions extracting dimensions from
\code{\link{grob}} objects.
}
\examples{
library(grid)
testCircle <- circleGrob(x = 15, y = 30, r = 15, default.unit = "mm")
grid.newpage()
grid.draw(testCircle)
simple_grob_height(testCircle)
simple_grob_width(testCircle)

simple_grob_height(testCircle, unitTo = "in")
simple_grob_width(testCircle, unitTo = "cm")
}
\author{
Johannes Graumann
}
\seealso{
\code{\link{multi_panel_figure}}
}

