% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-docs.R
\name{multipanelfigure-deprecated}
\alias{multipanelfigure-deprecated}
\title{Superceded objects in multipanelfigure}
\usage{
addPanel( figure, ... )
}
\arguments{
\item{figure}{Object of classes \code{multipanelfigure}/\code{\link[gtable]{gtable}}
as produced by \code{\link{multi_panel_figure}} and representing the figure the
panel is to be placed in.}

\item{...}{Arguments to functions you shouldn't use.}
}
\description{
Functions that are no longer used or have been superceded by
functions with underscore-separated names.
}
