% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{rhino}
\alias{rhino}
\title{A photo of a rhino}
\format{A JPEG file.}
\description{
A photograph of a greater one-horned rhinoceros (Rhinoceros unicornis) taken
in Kaziranga National Park, Assam, India.
CC-BY Janette Cotton, 2016.
}
\examples{
if(requireNamespace("magrittr"))
{
  `\%>\%` <- magrittr::`\%>\%`
  fig <- multi_panel_figure(
    width = 150, height = 150,
    rows = 3, columns = 3,
    interPanelSpacing = 0
  )
  rhinoFile <- system.file("extdata/rhino.jpg", package = "multipanelfigure")
  fig \%>\%
    add_panel(rhinoFile) \%>\%
    add_panel(rhinoFile, left_panel = 2, right_panel = 3) \%>\%
    add_panel(rhinoFile, top_panel = 2, bottom_panel = 3) \%>\%
    add_panel(rhinoFile, left_panel = 2, right_panel = 3, top_panel = 2, bottom_panel = 3)
}
}

