% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grobdimensions.R
\name{simplegrobwidth}
\alias{simplegrobheight}
\alias{simplegrobwidth}
\title{Convenient Access to \code{grob} Dimensions}
\usage{
simplegrobwidth(grob, unitTo = "mm")
simplegrobheight(grob, unitTo = "mm")
}
\arguments{
\item{grob}{A \code{\link[grid]{grob}} object for which dimensions are to be
retrieved.}

\item{unitTo}{A single \code{\link{character}} string representing a valid
\pkg{grid}-\code{\link[grid]{unit}}.}
}
\value{
Single \code{\link{numeric}} objects are returned.
}
\description{
Convenience functions extracting dimensions from
\code{\link{grob}} objects.
}
\examples{
library(grid)
testCircle <- circleGrob(x = 15, y = 30, r = 15, default.unit = "mm")
grid.newpage()
grid.draw(testCircle)
simplegrobheight(testCircle)
simplegrobwidth(testCircle)

simplegrobheight(testCircle, unitTo = "in")
simplegrobwidth(testCircle, unitTo = "cm")
}
\author{
Johannes Graumann
}
\seealso{
\code{\link{multipanelfigure}}
}

